Remove thanos-staging saturation rules from global recordings
This removes the `thanos-staging` saturation rules from the general
saturation rules that live in
`thanos(-staging)?-rules/autogenerated-saturation-(.*).yml` files.

These saturation rules are also in the
`thanos(-staging)?-rules/autogenerated-thanos-self-monitoring-saturation.yml`
file, so there was no need to duplicate them. Especially not including
the `thanos-staging` rules in the rules evaluated by production
thanos.

It also adds the correct `env` and `environment` static-labels to the
saturation resources for thanos-staging. These used to be set to
`thanos` but should be set to `thanos-staging` to be in line with the
SLI rules.