Fix labels for thanos-staging SLIs
These metrics don't get the correct type label everywhere on
ingestion, so let's ignore that for now and just rely on the
`namespace` label globally.

Some of the storegateway, ruler and compactor metrics also have
different formats than production. This fixes those to the correct
ones so we have some rules. We might want to update these to match
what we currently have in production when we deploy a new rules.

For https://gitlab.com/gitlab-com/gl-infra/reliability/-/issues/24423