Recursively apply recording rules in a directory
For Prometheus, this won't make a difference yet. As we don't have
rule-files nested in directories there.

For Thanos-staging we do, the experimental rules are currently
there.

This is a follow up from ba04090d707a21a791477dce6d2bfa12fcf8cbdd
which did not cover this part of getting the rules out.

Needed for
https://gitlab.com/gitlab-com/gl-infra/scalability/-/issues/2483