refactor(saturation): update gitaly active node ratio to 0.25
What
---
Update the ratio to be 0.25 instead of 0.20

Why
---
Part of the [multi-project migration](https://gitlab.com/gitlab-com/gl-infra/readiness/-/blob/master/library/gitaly-multi-project/README.md)
we are going to combine the [`marquee`
shard](https://gitlab.com/gitlab-com/gl-infra/readiness/-/blob/master/library/gitaly-multi-project/README.md#marquee-shard-gprd)
in the `default` shard. We'll do this at the end of the migraton, after
we've moved the `default` shard. To make sure that the marquee shards
get distributed evenly we will change the cut off to be at 75% so when
we are done we'll increase it back to 80% and distribute the marquee
shard evenly to the existing Gitaly storage.

Reference: https://gitlab.com/gitlab-com/gl-infra/reliability/-/issues/24553
Signed-off-by: Steve Xuereb <sxuereb@gitlab.com>