fix(alerts): ignore shards in sidekiq traffic cessation
What
---
Ignore the shards below for the trafficCessation alerts:

- `gitaly-throttled`
- `database-throttled`
- `quarantine`
- `urgent-authorized-projects`
- `elasticsearch`
- `memory-bound`

Why
---
In https://gitlab.com/gitlab-com/gl-infra/production/-/issues/16956 we
see `gitaly-throttled` cessation alert firing because we stopped running
a background migration of the Gitaly servers. There was nothing wrong
with the queue itself.

As we see in
https://gitlab.com/gitlab-com/runbooks/-/merge_requests/6373 we used to
be able to define `trafficCessationAlertConfig` per shard, and some of
them had it disabled, but in
https://gitlab.com/gitlab-com/runbooks/-/merge_requests/5654 we changed
how we define the SLI for the shard and didn't update the traffic
cessation alert.

Reference: https://gitlab.com/gitlab-com/gl-infra/reliability/-/issues/24555
Signed-off-by: Steve Xuereb <sxuereb@gitlab.com>