docs(gitaly): document how to move repo out of read-only
What
---
Add documentation on how to check if a repository is in read-only mode, and
how to make it writeable.

Why
---
Moving the storage server of a repository can fail, and if there is a
bug it could leave the repository in read-only mode, and we don't have a
documented way to fix it.

Signed-off-by: Steve Xuereb <sxuereb@gitlab.com>