Include all quantiles as templates for tamland
This makes sure we pass all the possible recorded quantiles used for
capacity planning on as templates for tamland.

We override the `quantile95_1h` one again to not use a recording, even
though it has one to avoid busting the tamland-data cache for those
queries. Ideally we'd also use the recording for that.

This is extracted from
https://gitlab.com/gitlab-com/runbooks/-/merge_requests/6469