Reinstate experimental rules with extraSourceSelector
Instead of passing in a modified aggregation set, we pass in the
selectorHash separately.

The `aggregationSet.selector` is otherwise used as a way to select
metrics from the aggregationSet: for example, the aggregations that we
used to record in Prometheus have a selector `monitor!="global"`: this
allows Thanos to avoid metrics that are recorded in Thanos already.

Passing in the selector separately will allow us to separate these
even more if we need to.

This also re-enables the experimental rules that were previously
disabled.

This is an alternative of
https://gitlab.com/gitlab-com/runbooks/-/merge_requests/6466 with a
fix for the global recording per service.