Update shard-detail dashboard
Instead of using the time of a shard, use the currently in use workers
like we do for the `sidekiq_shard_workers` component.

This allows us to remove some of these recording rules.

For duration we use the `sidekiq_job_completion_seconds_sum` metric
directly. But we collapse the panel by default as the high cardinality
makes it slow to load.