Remove transactional rates from experimental rules
This removes the transactional rates from experimental rules. As
discussed in
https://gitlab.com/gitlab-com/gl-infra/scalability/-/issues/2483#note_1639260872
these recordings are quite expensive, and we won't focus on them for
now.

We can still use these temporary aggregations to see if we can start
recording all of our aggregations from source metrics without the
intermediate rules in Prometheus. So I've renamed the file
accordingly. This won't work just yet, as the web rules rely on
intermediate recording rules that thanos-staging does not have. We'll
work on that in
https://gitlab.com/gitlab-com/gl-infra/scalability/-/issues/2602