Fix kube_container_cpu aggregation
In https://gitlab.com/gitlab-com/runbooks/-/merge_requests/6469 we
changed the `kube_container_cpu` saturation point, but accidentally
hardcoded the aggregation labels. In doing that, we forgot about the
`shard` label, which means that the metric was missing for capacity
planning for Sidekiq.

Part of https://gitlab.com/groups/gitlab-com/gl-infra/-/epics/946