Add 'type' filter to the runners dashboards
With the introduction of QA runners, which are using the same kind of
configuration but are not part of the SaaS Runners platform (which in
metrics and infra we call `ci-runners` service and type) and with them
being owned (most probably) by other people than SaaS Runners, the
runners are being added to Prometheus with a different `type`.

Instead of `ci-runners` it will be `qa-runners`.

This change makes sure that both types can be viewed through our set of
dashboards dedicated for GitLab Runner deployment maintenance.