Update "type" filtering in ci-runners dashboard
We've recently introduced a possibility to filter the ci-runners
dashboards with the `type` filter -
https://gitlab.com/gitlab-com/runbooks/-/merge_requests/6574.

This, however, used the default `type` filter which provides a ton of
types that are not usable for the ci-runners dashboards.

This change will make this filter to use only types that are related to
runner metrics.