refactor(sli): ignore ProjectExportWorker from sidekiq
What
---
Ignore the `ProjectExportWorker` from our SLI calculation for the `sidekiq` service.

Why
---
In https://gitlab.com/gitlab-com/gl-infra/production/-/issues/17165#note_1661925874 the p95 for the worker is always close to the threshold,
and the p99 always surpasses it.

This is a known issue and will be worked on at https://gitlab.com/groups/gitlab-org/-/epics/7940,
in the meantime, we shouldn't page the on-call for this worker.

Reference: https://gitlab.com/gitlab-com/gl-infra/production/-/issues/17165
Signed-off-by: Steve Xuereb <sxuereb@gitlab.com>