feat(saturation): add per engine inference limit
Anthropic does not only enforce a limit per model, but also a global
limit that is equal to the maximum number of concurrent requests to
the model that allows us the most requests.

We are currently using 2 models: `claude-2.0` and `claude-2.1`. We're
allowed more concurrent requests to `claude-2.1`, that is also our
global limit.

This adds a saturation point with alerts for that kind of saturation.

Part of https://gitlab.com/gitlab-com/runbooks/-/issues/143