chore: move anomaly detection rules to jsonnet
This migrates the anomaly detection rules from a regular yaml to
jsonnet. Doing this allows us to inject extra selectors. As a result
we can separate the rules by environment in Thanos.

In https://gitlab.com/groups/gitlab-com/gl-infra/-/epics/1107 we can
take this one step further and separate by service.

While doing this, we also remove the dependency on the 1m burn rate
that we intend to remove, and rely on the 5m burn rate and 1h burn
rate instead. The recording rule names remain the same. Doing this
using jsonnet and passing in the aggregation set should also help us
reusing this for the migration to Mimir.

Not using the 5m burnrate for these solves
https://gitlab.com/gitlab-com/gl-infra/platform/runway/team/-/issues/169