feat(saturation): Add kube_container_throttling
This adds a saturation point for `kube_container_throttling`. It
measures the number of periods per second a container is
throttled.

We're doing this over a period of 5m to filter out very short outliers
of bursty CPU utilization.

We're also taking the 99th quantile of this to filter out the worst
pods that might handling a particularly CPU intensive bursty workload.

We're currently targeting a 40% capacity planning SLO and a 50%
alerting SLO. This means that we want the 99th percentile of busiest
containers to be throttled less than 40% of the periods the container
is on CPU, over 5 minutes. In my opinion, this SLO is way too
lenient. But it is where some of our services currently peak (`web` is the main
offender here). I think we should move forward with these lenient SLOs
and revisit them in
https://gitlab.com/groups/gitlab-com/gl-infra/-/epics/946.

This is part of the work in
https://gitlab.com/groups/gitlab-com/gl-infra/-/epics/946

Discussed in
https://gitlab.com/gitlab-com/gl-infra/scalability/-/issues/2724#note_1723377089