feat: set generated attribute for autogenerated files
This sets gitlab-generated for files with for which the name matches
`autogenereted-*.yml`.

This will allow us to collapse the diff for this file by default when
we enable the `collapse_generated_diff_files` feature flag.

This feature flag is being enabled in
https://gitlab.com/gitlab-org/gitlab/-/issues/432670

We can enable it for this repo with

```
/chatops run feature set --project=gitlab-com/runbooks true
```