Make `supportsDetails` hidden
This change makes the `supportsDetails()` helper funtion hidden. This
is the function definition and we should never be attempting to render
this out.

This change makes it easier to manipulate and diagnose individual
metrics-catalog entries.

We can still refer to this function from other jsonnet as before with
this change.