Add the entry for istio.jsonnet to all.jsonnet
This change adds the entry for istio to the list in
`all.jsonnet`. This includes it in the generation pipelines.

Also included is a change to the `deploymentSelector`. This is
included as we need a map of labels to correctly select metric labels.

Looking in the `pre` cluster we can see that the `istiod` deployment
has the `app` label that we can use.

```
kubectl get deploy -n istio-system --show-labels istiod
NAME     READY   UP-TO-DATE   AVAILABLE   AGE    LABELS
istiod   2/2     2            2           244d   ...,app=istiod,...
```

This should select all of the metrics that we require.