fix: Robust cache invalidation using realpath
This addresses the problem outlined in https://gitlab.com/gitlab-com/runbooks/-/issues/149.

The intention is to expand and normalize the path given, so we always
use the same path. This is relevant as the path is part of the content
generated (see `warning`).

Before this change, the following calls would behave differently:

`make generate` -> uses ./path/to/some.jsonnet
`./scripts/generate-jsonnet-rules.sh path/to/some.jsonnet -> ends up
without trailing ./

Closes https://gitlab.com/gitlab-com/runbooks/-/issues/149