chore: remove obsolete commit-worker alerting rule
This was an alerting rule that triggered when the size of a critical
Sidekiq queue processing new pushes was not handling it's workload.

This queue does not exist anymore, instead we have a queue per Sidekiq
shard. This worker now runs on the `urgent_other` queue. So we haven't
seen this queue in a while, though the data structure is still
present.

Related to
https://gitlab.com/gitlab-com/gl-infra/scalability/-/issues/2584#note_1760601382