chore: remove high throttle rate alertRule
This removes the HighThrottleRate causal alert. This alert was being
ignored most of the time. It has been replaced by the
`kube_container_throttling` saturation point instead.

The alert was mainly firing for pods for `cloudflare-exporter` and
`clickhouse` both resources are missing the `type` label in
kubernetes. Which causes them to be missed by the saturation point.

We can handle the missing labels separately to make sure these
containers are also included in our saturation alerting.

Related to
https://gitlab.com/gitlab-com/gl-infra/scalability/-/issues/2584