feat: generate the component aggregation in Mimir
This is the first iteration of generating a component aggregationset
in Mimir. This should start populating the operation rates so they
would become visible on dashboards.

To achieve this, while keeping the existing implementation in
Prometheus and Thanos intact, we allow overriding the recording rule
configuration for Mimir. While keeping the defaults in the
`get-hybrid` and the current recording rule configuration. The global
default configuration lives in
`gitlabMetricsConfig.recordingRuleRegistry` and we default to the
`selectiveRegistry`.

The `selectiveRegistry` is the recording rule registry that only adds
recording rules for metrics specified in a `recordingRuleMetrics`
stanza on a service.

In `mimir-rules-jsonnet/mimir-aggregations.jsonnet`, the file
responsible for generating recording rules from source metrics into an
aggregation set. We specify the new `unifiedRegistry`.

The `unifiedRegistry` adds recording rules for all metrics used in
SLIs of a service, and it records this per service. This was added in
https://gitlab.com/gitlab-com/runbooks/-/merge_requests/6711 and
slightly refactored here to make it usable in the same way as the old
`selectiveRegistry`. The recording rules for the `selectiveRegistry`
aren't adjusted in this MR. We're only adding new ones for the
component aggregation.

Missing in this MR:
- Upscaling longer burn rates for aggregation sets, we'll do this in
https://gitlab.com/gitlab-com/gl-infra/scalability/-/issues/2898
- Recording ratios from self in the aggregation sets: the get-hybrid
  currently does this using
  [reflected rule
  groups](https://gitlab.com/gitlab-com/runbooks/blob/1e8922df6bc88868c404bf1b494a8bcc2afdc62f/reference-architectures/get-hybrid/src/rules/rules.jsonnet#L33). We'll
  do this in
  https://gitlab.com/gitlab-com/gl-infra/scalability/-/issues/2899
- Recording a component
  mapping (https://gitlab.com/gitlab-com/gl-infra/scalability/-/issues/2900)

This is part of https://gitlab.com/gitlab-com/gl-infra/scalability/-/issues/2475