fix: get-hybrid does not do shard/node monitoring
These aggregation sets are not defined on the get-hybrid config. There
are also no SLIs that configure node monitoring or shard monitoring in
the get-hybrid, once there are we'll get errors when passing around
the `null` values.

Discovered while working on
https://gitlab.com/gitlab-com/runbooks/-/merge_requests/6874. This is
visible in that merge request because we reach out to
`aggregationSet.upscaleBurnRate` inside the
`componentMetricsRuleSetGenerator` in that MR. That in turn caused us
to manifest the `shardAggregationSet` and `nodeAggregationSet`, which
don't exist in the get-hybrid architecture.