chore: remove unused burn rate recordings
This removes the 30m, 6h and 3d burn rates from the recording rules
used for error budgets for stage groups.

They are currently not used, because we don't have MWMBR alerting for
those aggregations. When we add
alerts (https://gitlab.com/groups/gitlab-com/gl-infra/-/epics/615)
we'll have to re-add these burn rates.

Removing them for now, this makes
https://gitlab.com/gitlab-com/runbooks/-/merge_requests/6874 a bit
easier to deal with as we don't need to handle upscaling aggregation
sets  upscaling in mimir right now...

We'll do it it in a separate issue. We'll still need to do it for
alerting in the project, but we can get the first aggregations
finished in
https://gitlab.com/gitlab-com/gl-infra/scalability/-/issues/2602