chore: remove the feature category ratios
This removes the errorRatio and apdex ratio from the feature category
aggregation. Since the feature category aggregation is an intermediate
aggregation set, we never use these ratios. We use this aggregation
set to aggregate further into the stage group aggregation.

To do this, we did need to stop generating the SLO analysis
dashboards. Those dashboards are used for validating when a MWMBR
alert would fire for an aggregation. But we don't alert on this
aggregation at all.

We do use the ratios from  the stage group aggregation we do use the
ratio recordings on the dashboards:
https://dashboards.gitlab.net/d/stage-groups-detail-source_code/adef71ed-25e3-5586-9d03-2a84baa1a862?orgId=1&from=now-7d&to=now

Discovered this while working on
https://gitlab.com/gitlab-com/gl-infra/scalability/-/issues/2899. Where
we would add more aggregations to the feature category aggregations
for globally evaluated services if we recorded a ratio there.