feat: split reflected ratios in component metrics
This records the reflected ratio for a component in a service inside
the rule group that also records the rates. This allows us to decouple
recording rules for services and components from each other, and it
uses the datapoints that were recorded before in the rule group.

When adding these recording rules, we use the static-labels together
with any other selectors passed as the selector for the ratio
recording. Any static labels are removed from the set of labels we
aggregate for.

For https://gitlab.com/gitlab-com/gl-infra/scalability/-/issues/2899