feat: Validate type selectors against emittedBy
Follow up from
https://gitlab.com/gitlab-com/runbooks/-/merge_requests/6880#note_1795883866
where we want to validate SLI's emittedBy needs to match its type
selector.

The `emittedBy` field is used to scope the SLI aggregations recording
rules with an individual `type=` selector per rule. Thus if the type
does not match the one in `emittedBy`, we would lose the recorded rule
for that particular type selector.

This validation is also useful to keep the `emittedBy` up to date in
case the SLI adds/removes a type selector.