chore: adjust runway apdex threshold for AI-gateway
This sets the runway satisfied threshold for requests to 27.2
seconds.

This runway threshold, using stackdriver metrics, encapsulates all of
the server type SLIs defined in the AI-gateway service. The largest
tolerated threshold defined there is the one for
`server_code_generations` and is set at 25s. So the runway SLI should
consider those.

The weird bucket values are coming from the stackdriver exporter. We
don't pick those ourselves.