feat: Record service component mapping rules in Mimir
Resolves https://gitlab.com/gitlab-com/gl-infra/scalability/-/issues/2900

The component mapping rules in the current Prometheus/Thanos setup is
currently generated in both Prometheus's and Thanos's
`service-key-metrics.jsonnet` which covers all services be it
`dangerouslyThanosEvaluated` or not.

We already have `service-slos.jsonnet` in both Mimir and
Thanos which also records the component mapping rules. The one in Thanos
excludes the `dangerouslyThanosEvaluated` services as they are already
recorded in thanos-rules/autogenerated-key-metrics-<service>.yml.

For Mimir, we can use the `service-slos.jsonnet` to generate the
component mappings for all services in a file. This is done by passing
the list of services for the component service mapping rules so that we
can pass all services for Mimir, while only passing the
non-`dangerouslyThanosEvaluated` for Thanos.