feat: generate service aggregation rules in mimir
This generates the service aggregation from the component aggregation.

To do this, we moved the selection of which aggregation set should
source a second level aggregation set into the second level
aggregation set itself using the `sourceAggregationSet` field. This
makes it clearer which aggregation sets are sourced from another
aggregation and which are from source metrics or SLI aggregations.

This will make the rest of the aggregations easier to define as well.

For https://gitlab.com/gitlab-com/gl-infra/scalability/-/issues/2475