refactor: source aggreation sets can list SLIs
This allows an aggregation set to list the SLIs that it is applicable
to. Some aggregations are only applicable to some SLIs or not at all
for a single service. This allows us to make that list using a
predicate in the aggregation set.

We can use this to detirmine if a service needs to be aggregated in a
set (for example for the `node-component` aggregation). We can also
change the SLIs that are going to be included in the aggregation (for
example for the `component-shard` aggregation).

This also moves the listing to the `aggregation-sets` file as hidden
fields, meaning we probably don't need to touch any other files when
we add aggregation sets.