feat: add the shardComponentSLIs aggregation
The component shard aggregation is used for monitoring services that
are deployed to different shards. For example Sidekiq, in the future
perhaps redis-sidekiq. This generates this aggregation so the sidekiq
dashboard should be populated.

This also removes the `intermediateSource: false` specification from
the nodeComponentSLIs, that is already part of the
`mimirAggregationSetDefaults`

For https://gitlab.com/gitlab-com/gl-infra/scalability/-/issues/2475