feat: separate SLI-aggregations into multiple groups
This separates the `sli_aggregation:` recording rule groups into
several groups. Before this change we had one group per burn rate per
service. After this change we have one group per burn rate per
SLI.

The metrics that are grouped together are the metrics that could be
used in a ratio calculation, so an SLI can have 2 groups:
- ops-rate & error rate
- all metrics used for apdex (histogram, success-rate, weight)

This makes it quicker to execute the rules sequentually in a group,
making it more likely that the group does not exceed it's interval
resulting in a missed execution causing a "gap" in the metrics.

Metrics that are used in multiple SLIs only get recorded once, in the
group they are "first" used.

Right now, each rule-group for a service deployed to kubernetes has a
rules for each cluster. Potentially, we could separate this out even
more so each cluster is evaluated in a group.

This is for
https://gitlab.com/gitlab-com/gl-infra/scalability/-/issues/2918