fix: avoid generating rule-files with 0 groups
We should avoid generating files with 0 groups rather than with 1
group.

This was an error pointed out in
https://gitlab.com/gitlab-com/runbooks/-/merge_requests/6981#note_1815065306. It
doesn't actually change anything ti to the recording rules because we
never have a file that would have less than 3 groups if there was an
SLI. We know this because we generate rules for 3 burn rates.