feat: add the node-service aggregation
This adds an aggregation set per node for a service. This aggregation
is used on the top in gitaly's node-dashboard.

https://dashboards.gitlab.net/d/gitaly-host-detail/gitaly3a-host-detail?from=now-6h%2Fm&to=now%2Fm&var-PROMETHEUS_DS=e58c2f51-20f8-4f4b-ad48-2968782ca7d6&var-environment=gprd&var-stage=main&orgId=1

Though the dashboard is currently empty because of a label mismatch:
https://gitlab.com/gitlab-com/gl-infra/scalability/-/issues/2700#note_1827690451.

To make it so we don't generate recording rules for services that we
don't need a node-aggregation for, we've added the `enabledForService`
predicate to the AggregationSet concept.

The last aggregation to tackle in https://gitlab.com/gitlab-com/gl-infra/scalability/-/issues/2475