chore: add location resource label for runway
This adds the `location` resource label for saturation metrics that
apply to Runway. This is required for viewing the resouce saturation
on separated by region in several dashboards.

For capacity planning the `max` or `histogram_quantile` across all
regions and locations will be used.

The `location` label is a workaround for the `region` getting
overwritten by the `external_labels` in the context of the
Thanos-sidecar. This problem does not occur when looking at these
metrics in Mimir because in that case the label that is already
present on the metrics takes precedence.

This is part of
https://gitlab.com/gitlab-com/gl-infra/scalability/-/issues/3392