chore: use only aggregation labels from mimir in unified registry
This makes it so we only use aggregation labels from aggregation sets
defined for Mimir for the unified recording rule registry.

Making sure that we're only aggregating what we need.

In practice, this means that we're removing the `queue`, `worker`
and `urgency` labels coming from the `sidekiqWorkerQueueSLIs`
aggregation set.

I noticed this when working on
https://gitlab.com/gitlab-com/gl-infra/platform/runway/team/-/issues/185,
for which it woulld be handy to be able to decouple Mimir rules and
Thanos rules.