feat: support location label in globally evaluated regional aggregation
This adds the `location` label to both regional aggregation sets. This
is required, because we use these aggregation sets to build the
regional dashboards. So including the label here will correctly
separate the series on the graph including the `location` label.

We need to support the `location` label because the Thanos-sidecar
overwrites the `region` label present on metrics with the
external label advertised by Prometheus. See
https://gitlab.com/gitlab-com/gl-infra/scalability/-/issues/3392#note_1852103750
for more information about this.

We will probably not need to add the `location` label in mimir, for
the metrics that are remote-written from stackdriver-exporter, this is
already correct. But we might still have to add it to mimir for
metrics collected from the regionally deployed application itself. We
will figure do that, if needed, in
https://gitlab.com/gitlab-com/gl-infra/scalability/-/issues/2722.

To support a globally evaluated, but regionally deployed service, we
generate the regional recording rules from source metrics in
Thanos. This is similar to what we already do in Mimir.

This is required for
https://gitlab.com/gitlab-com/gl-infra/platform/runway/team/-/issues/185