feat: add `location` to the legend on regional dashboards
This adds the `location` label to the legend for regional dashboards
as that is what is going to be used for globally evaluated services
using stackdriver metrics. For most other metrics this should not be
present, so it won't be visible on the dashboards.

We can revert this with
https://gitlab.com/gitlab-com/gl-infra/scalability/-/issues/3398

For https://gitlab.com/gitlab-com/gl-infra/platform/runway/team/-/issues/185