fix: remove location label from the service-regional aggregation
This is was over-eagerly introduced in
https://gitlab.com/gitlab-com/runbooks/-/merge_requests/7085. There we
started to record the `regionalComponentSLIs` with a `location`
label. We needed that to support the metrics coming from stackdriver
for globally evaluated services.

Label on the `regionalServiceSLIs` though, because that needs to have
a single label to aggregate both across `location` and `region`
otherwise we'll have multiple graphs for the same region: one with the
location populated for SLIs that use stackdriver metrics, and one with
`location=""`.

For
https://gitlab.com/gitlab-com/gl-infra/platform/runway/team/-/issues/185
and https://gitlab.com/gitlab-com/gl-infra/scalability/-/issues/3392