feat: record the service-regional aggregation for globally evaluated services
This aggregates the global services `gitlab_regional_sli_*` into
`gitlab_service_regional_*` metrics. In the Prometheus+Thanos world,
we aggregate both of these from the `gitlab_component_*` aggregation
sets using a `region` label from the external_labels advertised by
Prometheus.

To get a `region` label for globally evaluated metrics (like for
AI-gateway) we have to use the `location` or the `region` label. The
`region` label will also be populated by the `serviceMonitor`. But
some of the Runway SLIs come from `stackdriver-exporter` that is
scraped by another Prometheus in our main region. For those we need to
use the `location` label.

We do this by using the expensive [`label_replace`][0] PromQL-function, it
will override the existing `region` with the value from `location` if
that is present. If it isn't it will return the series untouched. This
way, we make the `location` override the value in `region` for the
`gitlab_service_regional_*` metrics.

This is part of
https://gitlab.com/gitlab-com/gl-infra/platform/runway/team/-/issues/185
and https://gitlab.com/gitlab-com/gl-infra/scalability/-/issues/3392

[0]: https://prometheus.io/docs/prometheus/latest/querying/functions/#label_replace