fix(runway): Only add regional labels when needed
This changes the `runwayHelper` to a function that will compute the
significant labels based on the service (self) passed. Right now it
only adds the `region`label if the service is deployed regionally.

It also adds the `location` label to the `runway_ingress` SLI so we
can also see per-region request rates in Thanos.

For
https://gitlab.com/gitlab-com/gl-infra/platform/runway/team/-/issues/185
and https://gitlab.com/gitlab-com/gl-infra/scalability/-/issues/3392