chore: rename waf service to cloudflare
This renames the `waf` service to `cloudflare`. This also matches the
`type` label we already have on the metrics.

It also specifies the right tenant for this service, metrics are only
available in the `gitlab-ops` tenant.

This will unblock further work for defining utilization metrics.

For https://gitlab.com/gitlab-com/gl-infra/scalability/-/issues/2917