feat: update inference SLI
This updates the inference SLI to the new metrics that cover all model
usages, not just the code suggestion ones.

The new metrics also include an error rate, and will be labelled with
a feature category.

To make the logging links on dashboards more relevant, this also
allows passing in an array of filters rather than just a `matches`
object. The filters can than use an `exists` filter instead.

Part of
https://gitlab.com/gitlab-org/modelops/applied-ml/code-suggestions/ai-assist/-/issues/441.