docs: yubikey setup instructions
Adds a bullet point around not forgetting to add the yubikey-agent.sock to your `~/.zshrc` (or equivalent file). This was prompted by the fact I'd stopped the yubikey-agent instructions at the setup phase, missed this step and took a while to figure out why the keys weren't being generated. 

Adding this step should hopefully stop someone else from fumbling over it in the future. :)