chore: disable traffic cessation alerts for runway non-us
The traffic we have in non-us regions is much lower. So it's not
unlikely to hit 0 sometimes even if an hour ago we still had enough
requests to meat the minimum RPS threshold. This was causing flapping
alerts.

This limits the traffic cessation alerts to only the US-region for
now.

Closes https://gitlab.com/gitlab-com/gl-infra/scalability/-/issues/3515